Just move `get/set_module_attr` to `model_utils.py`.
